Make it build on latest devkitARM
- UTF8-ify where possible
- Remove custom linkerscript
- Update Makefiles
  - devkitPro/nds-examples@6afa09b2054c9f47685514c32873b3905721c9ee
- Comment out akloader binaries
  - This will be reworked soon:tm:
- Eradicate sdidentify
  - AKRPG specific
- Eradicate libelm
- Eradicate save64m
- Eradicate file operations
- Eradicate libio*
- Eradicate crtsmall
- Fix paths for new root drive naming in latest libfat
- dsrom: fix type cast issue in homebrew check
- MAX_FILENAME_LENGTH -> PATH_MAX
- adapt directory listing operations to new dkP way
- timer: unstaticify _factor
- Remove all flashcart-specific bits
- fix type of cPopMenu::itemBelowPoint
- gbaloader: use updated vramSetPrimaryBanks function
- Move arm9-specific headers to arm9

+++ arm9/source/ui/popmenu.cpp
@@ -18,7 +18,7 @@
     along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-//à
+//ï¿½
 
 #include "ui.h"
 #include "popmenu.h"
@@ -92,7 +92,7 @@ void cPopMenu::draw()
 
 void cPopMenu::drawItems()
 {
-    // Ñ­»·»æÖÆitemÎÄ×Ö£¬Óö¼û selected ÎÄ×Ö¾ÍÏÈ»æÖÆÑ¡ÔñÌõ
+    // å¾ªç¯ç»˜åˆ¶itemæ–‡å­—ï¼Œé‡è§ selected æ–‡å­—å°±å…ˆç»˜åˆ¶é€‰æ‹©æ¡
     for( size_t i = 0; i < _items.size(); ++i )
     {
         s16 itemX = _position.x + _itemTopLeftPoint.x;
@@ -197,7 +197,7 @@ bool cPopMenu::processTouchMessage( const cTouchMessage & msg )
     return ret;
 }
 
-u32 cPopMenu::itemBelowPoint( const cPoint & pt )
+size_t cPopMenu::itemBelowPoint( const cPoint & pt )
 {
     cPoint menuPos(position().x + _barLeft, position().y + _itemTopLeftPoint.y - 2); cSize menuSize(barWidth(), _itemHeight * _items.size());
     cRect rect(menuPos, menuPos + menuSize);
